Ajustado endpoints de crud do usar e adicionado validações, e alguns nomes de variaveis.

+++ src/main/java/com/postech/fiap/fase1/domain/service/UserService.java
@@ -5,6 +5,7 @@ import com.postech.fiap.fase1.domain.assembler.UserAssembler;
 import com.postech.fiap.fase1.domain.dto.UserInputDTO;
 import com.postech.fiap.fase1.domain.model.User;
 import com.postech.fiap.fase1.domain.repository.UserRepository;
+import com.postech.fiap.fase1.domain.validator.UserValidator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -18,6 +19,7 @@ public class UserService {
 
     private final UserRepository userRepository;
     private final PasswordEncoder passwordEncoder;
+    private final UserValidator userValidator;
 
     @Transactional(readOnly = true)
     public User getOne(Long id) {
@@ -41,13 +43,27 @@ public class UserService {
     }
 
     public User create(UserInputDTO userInputDTO) {
+        userValidator.validateUserCreation(userInputDTO);
         User user = UserAssembler.inputToModel(userInputDTO, passwordEncoder.encode(userInputDTO.getPassword()));
         return userRepository.save(user);
     }
 
+    public User updateDetails(UserInputDTO userInputDTO) {
+        User user = findById(userInputDTO.getId());
+        userValidator.validateUserUpdateDetails(userInputDTO);
+        UserAssembler.updateUserDetails(user, userInputDTO);
+        return userRepository.save(user);
+    }
+    public User updatePassword(UserInputDTO userInputDTO) {
+        User user = findById(userInputDTO.getId());
+        userValidator.verifyUserPasswordAndConfirmation(userInputDTO);
+        user.setPassword(passwordEncoder.encode(userInputDTO.getPassword()));
+        return userRepository.save(user);
+    }
     @Transactional
     public void delete(Long idUser) {
         User user = findById(idUser);
+        userValidator.verifyUserLoggedIsAdminOrOwner(user.getId());
         userRepository.delete(user);
     }
 }
